Update Project Portfolio Page
The current Project Portfolio Page johndoe.md is a template for the AB3
project containing dummy information.

It needs to be updated to be contextualised to the needs of our own
project, Flagship.

Update the overview of template Project Portfolio Page johndoe.md to
reflect our Flagship project.

Rearrange the segments of the Project Portfolio Page for better
readability.