Add filter command and tests
There currently isn't a command that allows users to view a subset of data.

Implementing a filter command that allows users to filter by category will allow for better user experience and better viewing of data.

Let's implement the filter command to improve overall functionality and tests to ensure no bugs are introduced.