Update UG for filter command
Currently, the UG for filter command does not follow the format of the rest of the UG and is not very comprehensive.

Improving the readability of the UG as well as the overall harmony is essential in guiding users in using Flagship in the most efficient manner.

Let's improve the UG by giving more examples of the different formats of the filter command and ensure it is up-to-date.