Fix filter bug when create new entry
Currently, the internship list is not filtered properly when a new entry is created.

The internship list should be filtered properly to correspond to the filter status as displayed by the GUI to avoid confusion.

Let's make sure that the internship list is correctly filtered when a new entry is created with the create command.